refactor: split user admin routes
